sentiment_analysis/requirements.txt
connexion==3.0.6
Flask==2.2.5
grpcio==1.62.1
protobuf==5.26.1
psycopg2==2.9.9
